Changed: Experimental changes to better allow up to a total of 8gb of texture memory being used. Changed: Increase max memory slider value to 3984MB (new tested safe max) allowing a total of 8gb.
Note: The Viewer will not crash should it use more than 4gb memory for any of the two memory pools but it will flip over into 0 (rather than -2000~ and possibly crash), this will not only break the texture trashing feature and the memory display. Not an issue though as you'd have to use TextureLoadFullRes to even get close to this much memory usage and then the Viewer doesn't care anyway.

--HG--
branch : Black Dragon