Fixed: Shadows on almost all avatars missing almost completely under pretty common circumstances.
I'm unsure whether this might just be the whole alpha masking shadow feature working weird with something somewhere in my render pipeline but whatever it is it, as long as the texture type always has the last and thus highest priority over everything, it can and will flag items as alpha blending even if they are not compatible, in the last moment, resulting in them being skipped by the shadow pass. DONT DO THIS. Don't use the texture type as indicator for anything, we got Materials and as such there is an alpha mode that will ALWAYS Always always clearly dictate which type of alpha rendering we are going to use and shadows should absolutely respect this user choice otherwise we will end up with the choice (thus main rendering) and the shadow pass desyncing.

--HG--
branch : Black Dragon